RFC: Check implicit int <-> bool conversions
I don't like it. Other thoughts?

+++ src/base/unicode/confusables.cpp
@@ -94,7 +94,7 @@ int str_utf8_comp_confusable(const char *str1, const char *str2)
 		int ch2 = str_utf8_skeleton_next(&skel2);
 
 		if(ch1 == 0 || ch2 == 0)
-			return ch1 != ch2;
+			return static_cast<int>(ch1 != ch2);
 
 		if(ch1 != ch2)
 			return 1;
